requirements.txt
# Arquivo de dependências do Python
#source venv/bin/activate

pandas==2.2.2
openai==1.30.1
python-dotenv==1.0.1